Combine reservations redux file with user-sessions
